Add admin files for info app
