+++ apps/info/admin/__init__.py
@@ -1,0 +1,3 @@
+from apps.info.admin.festival import ArtTeamMemberAdmin, FestivalAdmin, FestTeamMemberAdmin, PressReleaseAdmin
+from apps.info.admin.people import PartnerAdmin, PersonAdmin, SelectorAdmin, SponsorAdmin, VolunteerAdmin
+from apps.info.admin.place import PlaceAdmin
